+++ src/Exception/NotFoundException.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Phlib\LitmusResellerSDK\Exception;
+
+/**
+ * @package Phlib\Litmus-Reseller-SDK
+ */
+class NotFoundException extends \RuntimeException implements Exception
+{
+}